Version 3.28.28 (based on bleeding_edge revision r22455)
Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@22457 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arm/assembler-arm.cc
@@ -3685,7 +3685,8 @@ void ConstantPoolBuilder::Populate(Assembler* assm,
       if (type == ConstantPoolArray::INT64) {
         constant_pool->set_at_offset(offset, rinfo.data64());
       } else if (type == ConstantPoolArray::INT32) {
-        constant_pool->set_at_offset(offset, rinfo.data());
+        constant_pool->set_at_offset(offset,
+                                     static_cast<int32_t>(rinfo.data()));
       } else if (type == ConstantPoolArray::CODE_PTR) {
         constant_pool->set_at_offset(offset,
                                      reinterpret_cast<Address>(rinfo.data()));
